web/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
asyncpg==0.30.0
aio-pika==9.5.5
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
cryptography==44.0.2
dill==0.3.9
ecdsa==0.19.1
greenlet==3.1.1
inquirerpy==0.3.4
Mako==1.3.9
MarkupSafe==3.0.2
multiprocess==0.70.14
orjson==3.9.15
passlib==1.7.4
pfzy==0.3.4
prompt_toolkit==3.0.50
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
pyasn1==0.4.8
pycparser==2.22
python-jose==3.4.0
psycopg2; platform_system=="Windows"
psycopg2-binary; platform_system!="Windows"
robyn==0.65.0
rustimport==1.3.4
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.40
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.1
watchdog==4.0.1
wcwidth==0.2.13
nest_asyncio==1.6.0
